Add color_capabilities to light model Break out light tests
