+++ tests/lights/__init__.py
@@ -1,0 +1,4 @@
+"""Tests for pydeCONZ lights.
+
+pytest --cov-report term-missing --cov=pydeconz tests/lights
+"""
